Move auth provider into auth package
