+++ src/main/scala/com/sbuslab/sbus/rabbitmq/RabbitMqTransport.scala
@@ -23,7 +23,8 @@ import org.slf4j.{LoggerFactory, MDC}
 
 import com.sbuslab.model._
 import com.sbuslab.model.scheduler.ScheduleCommand
-import com.sbuslab.sbus.{AuthProvider, Context, Headers, Transport}
+import com.sbuslab.sbus.{Context, Headers, Transport}
+import com.sbuslab.sbus.auth.AuthProvider
 
 
 class RabbitMqTransport(conf: Config, authProvider: AuthProvider, actorSystem: ActorSystem, mapper: ObjectMapper) extends Transport {
